move data reader implementations to reader folder
