+++ elasticdl/python/data/reader/data_reader.py
