+++ elasticdl/python/data/reader/data_reader_factory.py
@@ -1,10 +1,10 @@
 import os
 
 from elasticdl.python.common.constants import ODPSConfig
-from elasticdl.python.data.csv_reader import CSVDataReader
 from elasticdl.python.data.odps_io import is_odps_configured
-from elasticdl.python.data.odps_reader import ODPSDataReader
-from elasticdl.python.data.recordio_reader import RecordIODataReader
+from elasticdl.python.data.reader.csv_reader import CSVDataReader
+from elasticdl.python.data.reader.odps_reader import ODPSDataReader
+from elasticdl.python.data.reader.recordio_reader import RecordIODataReader
 
 
 def create_data_reader(data_origin, records_per_task=None, **kwargs):
